+++ src/Trab2/Grafo.h
@@ -55,7 +55,7 @@ class Grafo{
     
     void Guloso(string instancia);
     int gulosoRandomizado(float alfa, int numInter, double tempo[], int seed);
-    void gulosoRandomizadoReativo(int numInter,float tempo[], int bloco, int seed, string instancia);
+    void gulosoRandomizadoReativo(int numInter,double tempo[], int bloco, int seed, string instancia);
     float escolheAlfa(int probAlfa[], int seed);
     void atualizaProb(int *probAlfa,int **resultBloco);
     void resetBloco(int **resultBloco, int bloco);
